+++ js/create-photos.js
