+++ js/filter-modal.js
@@ -5,67 +5,52 @@ import {resetScale} from './scale.js';
 import{pristineReset,validate} from './validation.js';
 import {sendData} from './api.js';
 
-//время задержки сообщения об ошибке при получении данных с сервера
 const ALERT_SHOW_TIME = 5000;
-//темплейт разметки об успешной загрузке изображения на сервер
 const successTemplate = document.querySelector('#success').content.querySelector('.success');
-//Копия темплейта об успешн загр изображения на сервер
 const successMessage = successTemplate.cloneNode(true);
-//Кнопка для отправки данных на сервер - ОПУБЛИКОВАТЬ
 const submitBtn = document.querySelector('.img-upload__submit');
-//Поведение кнопки ОПУБЛИКОВАТЬ
+
 const SubmitButtonText = {
   IDLE: 'Опубликовать',
   SENDING: 'Публикую...'
 };
-//красный значок инстаграмма на основном окне
+
 const imgUploadInput = document.querySelector('.img-upload__input');
-//модальное окно с фильтрами
 const imgUploadOverlay = document.querySelector('.img-upload__overlay');
-//большое фото
 const imgUploadPreview = document.querySelector('.img-upload__preview img');
+
 const thumbnails = [...document.querySelectorAll('.effects__preview')];
-// кнопка закрытия модального окна с фильтрами
 const imgUploadBtnCancel = document.querySelector('.img-upload__cancel');
-//ul,в котором радио-кнопки превью - фильтры
 const effectsList = document.querySelector('.effects__list');
-//инпут слайдера под ползунком
 const effectLevelInput = document.querySelector('.effect-level__value');
-//контейнер, в котором лежит слайдер
+
 const sliderContainer = document.querySelector('.effect-level__slider');
-//поле Изменения глубины эффекта, накладываемого на изображение
 const imgUploadEffectLevel = document.querySelector('.img-upload__effect-level');
 const form = document.querySelector('.img-upload__form');
 const errorLoadTemplate = document.querySelector('#error').content.querySelector('.error');
-//СООБЩЕНИЕ ОШИБКИ ЗАГРУЗКИ ПРИНЯТНЫХ ФОТО
+
 const dataErrorTemplate = document.querySelector('#data-error').content.querySelector('.data-error');
 const errorMessage = dataErrorTemplate.cloneNode(true);
 
-//ДЛЯ ПОДСТАНОВКИ ФОТО ПОЛЬЗОВАТЕЛЯ
 const FILE_TYPES = ['jpg', 'jpeg', 'png'];
 
-//сброс фильтра с главной картинки
 const resetFilter = () => {
   imgUploadPreview.style.filter = 'none';
 };
 
-//Обновление настроек слайдера для каждого эффекта
 const updateSlider = (evt) => {
-  //const currentInput = evt.target.closest('.effects__item');
   if (evt.target.classList.contains('effects__radio')) {
     const currentEffect = evt.target.value;
     sliderContainer.noUiSlider.updateOptions(Effects[currentEffect]);
   }
 };
 
-//Нажатие escape на модальное окно
 const onDocumentKeyDown = (evt) => {
   if (isEscapeKey(evt)) {
     closeUploadModal ();
   }
 };
 
-//Закрытие окна. Удаление обработчика escape. Удаление обработчика клика,Удаление посл класса с больш фото. удаление scale
 function closeUploadModal () {
   body.classList.remove('modal-open');
   imgUploadOverlay.classList.add('hidden');
@@ -79,10 +64,8 @@ function closeUploadModal () {
   imgUploadBtnCancel.removeEventListener('click', closeUploadModal);
 }
 
-//дефолтное значение спрятанного инпута
 effectLevelInput.value = 100;
 
-//Создала первичные настройки для слайдера
 noUiSlider.create(sliderContainer, {
   start: 100,
   step: 1,
@@ -105,21 +88,17 @@ noUiSlider.create(sliderContainer, {
 });
 
 sliderContainer.noUiSlider.on('update',() => {
-  //запись в спрятанный инпут значение ползунка
   effectLevelInput.value = sliderContainer.noUiSlider.get();
   const currentEffect = document.querySelector('.effects__radio:checked');
   getEffectToPhoto(currentEffect.value, effectLevelInput.value);
 });
 
-//Открытие мод окна. Добавление обработчика закрытия escape. Добавл обработчика Х
 const openUploadModal = () => {
-  //УБРАТЬ ЗДЕСЬ И В CLOSE В ОТДЕЛЬНУЮ ФУНКЦИЮ
   imgUploadOverlay.classList.remove('hidden');
   body.classList.add('modal-open');
   imgUploadEffectLevel.classList.add('hidden');
   document.addEventListener('keydown', onDocumentKeyDown);
   imgUploadBtnCancel.addEventListener('click', closeUploadModal);
-  //Ставлю слушатель на ul effects__list
   effectsList.addEventListener('change', updateSlider);
 };
 
@@ -136,7 +115,6 @@ const showFilterModal = () => {
   }
 };
 
-//На появление в инпуте файла, показываю модальное окно с котенком
 imgUploadInput.addEventListener('change',showFilterModal);
 
 function getEffectToPhoto (effect,value) {
@@ -148,7 +126,6 @@ function getEffectToPhoto (effect,value) {
   }
 }
 
-//Поведение кнопки опубликовать во время отправки
 export const blockSubmitBtn = () => {
   submitBtn.disabled = true;
   submitBtn.textContent = SubmitButtonText.SENDING;
@@ -169,16 +146,13 @@ const showDataErrorMessage = () => {
 
 const onSuccessButtonKeyDown = (evt) => {
   if (isEscapeKey(evt)) {
-    evt.preventDefault();//нужен
+    evt.preventDefault();
     closeSuccessMessage();
   }
 };
 
-//искать кнопку здесь или вне?
 export const showSuccessMessage = () => {
-  // Добавляю визуал об успехе
   document.body.append(successMessage);
-  //ставлю слушатель на кнопку закрытия
   const successOverlay = document.querySelector('.success');
   successOverlay.addEventListener ('click',closeSuccessOverlay);
   const successButton = successMessage.querySelector('.success__button');
@@ -188,10 +162,7 @@ export const showSuccessMessage = () => {
 
 function closeSuccessMessage() {
   successMessage.remove();
-  //successButton.removeEventListener('click', closeSuccessMessage); Как удалить обработчик клика, если кнопку ищу,когда окно открывается
   document.removeEventListener('keydown', onSuccessButtonKeyDown);
-  //Удаление по клику на кнопку. ЗАново искать successButton?
-  // successButton.addEventListener('click', closeSuccessMessage);
 }
 
 function closeSuccessOverlay (evt) {
@@ -200,16 +171,12 @@ function closeSuccessOverlay (evt) {
   }
 }
 
-//удаление визуала об ошибке
-// кнопка errorButtonClose ищется и существует только в showErrorMessage
 const closeErrorMessage = () => {
   errorLoadTemplate.remove();
-  //добавляю закрытие по escape всего мод окна
   document.addEventListener('keydown', onDocumentKeyDown);
   document.removeEventListener('keydown', onErrorButtonKeyDown);
 };
 
-//Функция для закрытия сообщения - ошибки по клику на шторку
 function closeErrorOverlay (evt) {
   if (evt.target.classList.contains('error')) {
     closeErrorMessage();
@@ -218,12 +185,11 @@ function closeErrorOverlay (evt) {
 
 function onErrorButtonKeyDown (evt) {
   if (isEscapeKey(evt)) {
-    evt.preventDefault();//нужен
+    evt.preventDefault();
     closeErrorMessage();
   }
 }
 
-//Сообщение ОШИБКА ОТПРАВКИ
 export const showErrorMessage = () => {
   document.body.append(errorLoadTemplate);
   const errorButtonClose = document.querySelector('.error__button');
@@ -235,7 +201,6 @@ export const showErrorMessage = () => {
   document.addEventListener('keydown', onErrorButtonKeyDown);
 };
 
-//Отправка формы, если форма проходит валидацию!
 form.addEventListener('submit', (evt) => {
   //если не поставить, то форма улетит на сервер
   evt.preventDefault();
