+++ js/render-comments.js
@@ -7,7 +7,6 @@ import {
 let comments = [];
 let currentCount = 0;
 const STEP = 5;
-//let COUNT = 0;
 
 const renderNextComments = () => {
   const fragment = document.createDocumentFragment();
@@ -34,7 +33,6 @@ const renderNextComments = () => {
 
 const clearComments = () => {
   currentCount = 0;
-  //COUNT = 0;
   commentsContainer.innerHTML = '';
   commentsLoader.classList.remove('hidden');
   commentsLoader.removeEventListener('click', renderNextComments);
