+++ js/scale.js
@@ -1,14 +1,14 @@
 import { imgUploadPreview } from './filter-modal.js';
 
-//уменьшение размера фотографии
+
 const scaleSmallerBtn = document.querySelector('.scale__control--smaller');
 const scaleBiggerBtn = document.querySelector('.scale__control--bigger');
 const scaleInput = document.querySelector('.scale__control--value');
+const MAX_SCALE_VALUE = 100;
+const SCALE_STEP = 25;
 
-//- 100% +
-//сброс scale
 const resetScale = () => {
-  scaleInput.value = `${100}%`;
+  scaleInput.value = `${MAX_SCALE_VALUE}%`;
   imgUploadPreview.style.transform = 'none';
 };
 
@@ -20,13 +20,13 @@ const renderScale = (value) => {
 
 scaleSmallerBtn.addEventListener('click', () => {
   const value = parseInt(scaleInput.value, 10);
-  const newValue = (value > 25) ? value - 25 : value;
+  const newValue = (value > SCALE_STEP) ? value - SCALE_STEP : value;
   renderScale(newValue);
 });
 
 scaleBiggerBtn.addEventListener('click', () => {
   const value = parseInt(scaleInput.value, 10);
-  const newValue = (value < 100) ? value + 25 : value;
+  const newValue = (value < MAX_SCALE_VALUE) ? value + SCALE_STEP : value;
   renderScale(newValue);
 });
