+++ js/utils.js
@@ -4,7 +4,7 @@ function getRandomInteger (min,max) {
   const result = Math.random() * (upper - lower + 1) + lower;
   return Math.floor(result);
 }
-//Количество фотографий
+
 const generateCountOfPhotos = () => {
   let count = 0;
   return function () {
@@ -12,7 +12,6 @@ const generateCountOfPhotos = () => {
   };
 };
 
-//создать уникальный id
 const createUniqueId = (min,max) => {
   const previousValues = [];
   return function () {
@@ -30,23 +29,14 @@ const createUniqueId = (min,max) => {
 
 const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
 
-//Проверка ESCAPE, экспортируем в ?main.js
 const isEscapeKey = (evt) => evt.key === 'Escape';
 
 function debounce (callback, timeoutDelay = 500) {
-  // Используем замыкания, чтобы id таймаута у нас навсегда приклеился
-  // к возвращаемой функции с setTimeout, тогда мы его сможем перезаписывать
   let timeoutId;
 
   return (...rest) => {
-    // Перед каждым новым вызовом удаляем предыдущий таймаут,
-    // чтобы они не накапливались
     clearTimeout(timeoutId);
-
-    // Затем устанавливаем новый таймаут с вызовом колбэка на ту же задержку
     timeoutId = setTimeout(() => callback.apply(this, rest), timeoutDelay);
-    // Таким образом цикл «поставить таймаут - удалить таймаут» будет выполняться,
-    // пока действие совершается чаще, чем переданная задержка timeoutDelay
   };
 }
